Refactored verfication code, added comments
Updated verification code to use the /status endpoint instead of a separate verify endpoint.
Updated ping to send the client info with every ping, which includes the client's mac addresses, and nickname.
This allows for future implementation of the server checking the last time a client sent a ping, and sending a warning if there's not a ping for a long time.